+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Add yaml resource type
+- Add makefile resource type
 
 ## 0.1.0 - 2019-03-07
 ### Added